move things to views
