+++ common/src/api/external/mod.rs
@@ -936,21 +936,6 @@ impl DiskState {
 }
 
 /*
- * SNAPSHOTS
- */
-
-/// Client view of a Snapshot
-#[derive(ObjectIdentity, Clone, Debug, Deserialize, Serialize, JsonSchema)]
-pub struct Snapshot {
-    #[serde(flatten)]
-    pub identity: IdentityMetadata,
-
-    pub project_id: Uuid,
-    pub disk_id: Uuid,
-    pub size: ByteCount,
-}
-
-/*
  * Sagas
  *
  * These are currently only intended for observability by developers.  We will
